Update rayleigh_1.m
performed Rayleigh integral for a baffled circular piston , mesh size can be adjusted from the script, Number of meshes should approximately equals to M_size^2 for the meshing of baffled circular piston and P_size^2 for the meshing of circular phase plate.
Takes to long to run: ->I am thinking maybe find functions to put matrix in the GPU? perhaps it will be faster.
                                  ->Or to reduce the mesh size
I think we can sanity check whether the intergral result is sound or not? before we proceed?